Fix mess after merging the main branch
There was a lot of consolidation to do. The big one being that the
packer/plugin package now needs to be part of packer core to avoid a
circular dependency issue. Still, I think it makes sense that the plugin
related types are so tyghly coupled to packer core. Especially in this
PR because the configuration now tells wether a plugin will be loaded.