Replace HashiBot with GitHub action for remote-plugin issue migration
This changes removes the use of HashiBot for migrating issues labels
with remote-plugin/* to their respective plugin repositories. The GitHub
action has feature parity with HashiBot so the migration should work as
it did before.

The GitHub action closes the original issues and creates a new issue on
the remote plugin repository with the original contents of the issue to
preserve history. We chose not to go with GitHub's issue transfer option
because it completely moves the issue to the target repo, meaning there
is not history that it was ever created on the original repository.

In the future, we will look to add an action for closing remote-plugin
issues that we no longer have access to.

Closes: #11043