Fix panic when running a Packer registry build in a clean directory
Give the case where a new build was executed using a template in a
clean non-Git controlled directory without specifying the
HCP_PACKER_BUILD_FINGERPRINT env. variable would result in a panic.

This was do to the call the NewIteration throwing an error that was not
being captured upstream when creating the bucket.

Results of tests before change
```
=== RUN   TestNewIteration/using_empty_git_directory
    types.iterations_test.go:94: expected "using empty git directory" to return with no error, but it Packer encountered an issue reading the git info for the path "/var/folders/_9/_q1cqjtd3358p1ldhy7j399m0000gn/T/empty-init".
        If your Packer template is not in a git repo, please add a unique template fingerprint using the env var HCP_PACKER_BUILD_FINGERPRINT. Error: reference not found
--- FAIL: TestNewIteration (0.65s)
```

Results of tests after change
```
--- PASS: TestNewIteration (0.71s)
    --- PASS: TestNewIteration/using_fingerprint_env_variable (0.00s)
    --- PASS: TestNewIteration/using_git_fingerprint (0.70s)
    --- PASS: TestNewIteration/using_empty_git_directory (0.00s)
    --- PASS: TestNewIteration/using_no_fingerprint_in_clean_directory (0.00s)
PASS
```