Update registry builder to skip a build that was found to be DONE
This is another fix for Packer trying to update a build that is already complete.

Update iteration complete check

This change fixes an issue where a completed iteration is allowed to be
updated with new builds after it has been given a version.

Reproduction steps.
- Run a single Packer registry build - upon success the iteration will be
given a version.
- Add a new source build to the build template; re-run the build.

-- Expected
Packer should fail, alerting the user that the iteration for the given
fingerprint is complete.

-- Actual
The build executes successfully and the iteration is updated with new
build image that were not part of the original version. Note this is
different from running a build with the only and except flags.

* Remove the publishing of post-processor metadata from the registry
  post-processor.