[CI] Add publish-release step to nightly release workflow
In reviewing the results of the nightly release workflow it was found
that upon creating the release it was successfully marked as a
prerelease. But it still needed to be manually publish as an official
prerelease. This change adds a new action to publish the created
release, as the last step.