Update community note on GitHub issue templates
In an effort to help prioritize issues (i.e bugs and feature requests) that are important to the community
the community note pertaining to issue voting has been moved to the top of each
template with a comment informing users not to delete it. The goal is to
rely more on votes for feature requests and issues that are not critical.

Critical issues being crashes, regressions, security vulnerabilities, and broken features.