Bump packer-plugin-sdk (#11754)
The Packer Plugin SDK has been updated in v0.2.12 to address the
x/crypto/ssh fix for unsupported key algorithms. This change is specific
to the SSH communication which is mostly in the sdk pkg. But it is being
brought into Packer to have parity with all the external plugins, which
are being updated separately.