packer: enable implicit required plugins
When running a HCL2 build, Packer can load plugins two ways: either from
a plugin binary that is locally installed, or from one of the bundled
plugins.

Since bundled plugins are on the way out, we are enabling implicitely
required plugins for HCL2 configurations.
This effectively means that if a plugin is not installed, we will point
users towards using `packer init' to install them, and if no
`required_plugins' block is specified, for the few official plugins we
support, we will automatically install the latest version of the plugin
locally.