packer: move BundledPluginsStatus to PluginConfig
The BundledPluginsStatus map keeps track of the bundled plugin usage in
a Packer template.

This was stored as a global map in the command package, but since we
want to use it elsewhere in the code, and because of some circular
dependencies, we opt to move this into the `PluginConfig' structure.