datasources: move evaluation to unit method
Datasource evaluation was embedded in the sequential scheduler for HCL
templates, but this has to move so we can individually invoke it later
on when we add the DAG to the codebase.

Therefore, we move the logic for evaluating a single datasource to the
PackerConfig once more, and call this evaluation function from the
scheduler.